attributes_base
