+++ spec/csv_row_model/concerns/model/attributes_spec.rb
@@ -1,72 +1,6 @@
 require 'spec_helper'
 
 describe CsvRowModel::Model::Attributes do
-  describe "instance" do
-    let(:options) { {} }
-    let(:instance) { BasicRowModel.new(options) }
-
-    describe "#attributes" do
-      subject { instance.attributes }
-      it "returns an empty hash" do
-        expect(subject).to eql(string1: nil, string2: nil)
-      end
-
-      context "with methods defined" do
-        before do
-          instance.define_singleton_method(:string1) { "haha" }
-          instance.define_singleton_method(:string2) { "baka" }
-        end
-
-        it "returns the map of column_name => public_send(column_name)" do
-          expect(subject).to eql( string1: "haha", string2: "baka" )
-        end
-      end
-
-      context "with nil returned in method" do
-        before do
-          instance.define_singleton_method(:string1) { nil }
-          instance.define_singleton_method(:string2) { "baka" }
-        end
-
-        it "returns the map of column_name => public_send(column_name)" do
-          expect(subject).to eql(string1: nil, string2: "baka")
-        end
-      end
-
-      context "with one method defined" do
-        before do
-          instance.define_singleton_method(:string1) { "haha" }
-        end
-
-        it "returns the map of column_name => public_send(column_name)" do
-          expect(subject).to eql(string1: "haha", string2: nil)
-        end
-      end
-    end
-
-    context "with attribute_objects stubbed" do
-      include_context "stub_attribute_objects", string1: "haha", string2: "baka"
-
-      describe "#original_attributes" do
-        subject { instance.original_attributes }
-
-        it "returns the attributes hash" do
-          expect(subject).to eql(string1: 'haha', string2: 'baka')
-        end
-      end
-
-      describe "#original_attribute" do
-        it_behaves_like "attribute_object_value", :original_attribute, :value, string1: "haha"
-      end
-    end
-
-    describe "#to_json" do
-      it "returns the attributes json" do
-        expect(instance.to_json).to eql(instance.attributes.to_json)
-      end
-    end
-  end
-
   describe "class" do
     let(:klass) { BasicRowModel }
