+++ spec/csv_row_model/concerns/model/base_spec.rb
@@ -32,27 +32,5 @@ describe CsvRowModel::Model::Base do
       subject { instance.abort? }
       it "never aborts" do expect(subject).to eql false end
     end
-
-    context "with attributes stubbed" do
-      let(:attributes) { { string1: "a", string2: "b" } }
-      before do
-        allow_any_instance_of(BasicRowModel).to receive(:string1) { "a" }
-        allow_any_instance_of(BasicRowModel).to receive(:string2) { "b" }
-        expect(BasicRowModel.new.attributes).to eql attributes
-      end
-
-      describe "#eql?" do
-        it "removes duplicate entries" do
-          expect([BasicRowModel.new,BasicRowModel.new].uniq.size).to eql(1)
-        end
-      end
-
-      describe "#hash" do
-        subject { instance.hash }
-        it "is the attributes hash" do
-          expect(subject).to eql attributes.hash
-        end
-      end
-    end
   end
 end
